Splice constants in strings infer proof cons (#11593)
This is in preparation for significant simplifications to the core rules
of the strings calculus.

This makes the proof elaborator for strings inferences perform the
appropriate "constant splicing" in preparation for applying string
rules.

For example, currently the checker ProofRule::CONCAT_EQ can infer
(= (str.++ "AB" x) (str.++ "A" y)) => (= (str.++ "B" x) y)
However, noting it is simple to prove (= "AB" (str.++ "A" "B")), proof
elaboration can simplify this to:
(= (str.++ "A" "B" x) (str.++ "A" y)) => (= (str.++ "B" x) y)
where notice that the checker for ProofRule::CONCAT_EQ now only has to
drop common arguments from the concatenation.

This PR does not yet change the internal proof checker or the Eunoia
definitions. Instead the proof constructor in strings only uses a subset
of behaviors allowed by the checker.