Simplify quantifiers variable elimination inequality rewrite (#11605)
This is work towards giving proofs for eliminating variables from
quantified formulas based on inequality reasoning.

Previously we computed a term substitution for all bounds on a variable;
now we simply drop those bounds from the body of the quantified formula
(noting that NNF always ensures the bodies of quantified formulas are
disjunctive only).

This impacts the definition of MACRO_QUANT_VAR_ELIM_INEQ, which will be
elaborated in a followup PR.