Simplifications to the core rules of the strings calculus (#11594)
The core rules of the string calculus are overcomplicated by making the
proof checker "splice constants". This can be simplified away at
elaboration time at very little additional cost, with the benefit of
making the definitions of the rules significantly simpler.

This PR finalizes this change in the Eunoia signatures (based on the
simplifications in https://github.com/cvc5/cvc5/pull/11628).

This is also work towards fixing the Eunoia signature for sequences,
which `$str_from_flat_form` does not work.