Do not purify ground Boolean subterms in quantifier bodies (#11654)
Fixes https://github.com/cvc5/cvc5-projects/issues/763.

More generally, this issue suggests it is not safe to introduce
purification skolems for any Boolean term, as this can make them
allocated as an "internal-only" SAT literal, where if later the same
purification skolem is introduced as a Boolean term skolem, the theory
solvers will not be notified. This will require more thought.