Add getGeneralizedConstRegExp utility in RegExp entail (#11608)
Adds 2 macro rewrite rules that rely on this, updates the string
rewriter to use the new utility in these 2 places. Note that in both
cases the rewrites now applicable in these cases are more powerful.

The elaboration for these rules will be added in a followup PR, and
depends on https://github.com/cvc5/cvc5/pull/11600.

My dev branch showed no performance regression on SMT-LIB and an
industrial set with this change.