Eliminate use of LinearRewriteStrategy in cases where proofs can not be elaborated (#11656)
This is work towards filling the remaining BV rewrite holes.

The use of `LinearRewriteStrategy` in the BV rewriter may severely
complicate our ability to infer proofs for rewrite rules. In particular,
since the linear rewrite strategy applies potentially multiple rewrites
at once, the rewrite reported to the main rewriter may be impossible to
infer.

For example we may rewrite `a -> b` by `R1`, `b -> c` by `R2`, and the
rewriter is given `a -> c` as a rewrite which does not fit the expected
pattern of `R1` or `R2`.

This PR also modifies the BV rewriter not to apply multiple rewrites at
once in cases where it makes a difference. That is, if `a->b`, then we
immediately return the rewrite.

It also makes two rewrites (ExtractExtract and BvMult2) only applicable
at *post* rewrite, which enables RARE to handle all cases where they
apply.

I tested this on all BV logics of SMT-LIB, where this PR was +120-120,
all benchmarks close to timeout, with neglible <0.01% difference in time
on commonly solved.