Improve recursion limit heuristic in RARE (#11683)
This addresses a major inefficiency in the RARE proof reconstruction.

In particular, some of the "builtin" tactics did not count against the
recursion limit. This meant that a simple rewrite like:
```
(bvult t 0) ---> false
```
could timeout if `t` was an enormous term, even though a user rule
matches this rewrite verbatim.

We now decrement the recursion limit in all cases, apart from two
tactics (cong and true_intro) which are benign.

---------

Co-authored-by: Abdalrhman Mohamed <abdoo8080@outlook.com>