Address nondeterminism in quantifiers rewriter (#11700)
This addresses a non-determinism in the quantifiers rewriter that leads
to (very rare) instances of proof checking failures.

The issue is caused by constructing a quantified formula with a literal
`(= (not x) (not y))`, where `x -> y` is a variable elimination.
Assuming this literal is garbage collected and then reconstructed, we
may have `(= (not y) (not x))` where `y -> x` is a variable elimination
later, thereby leading to a very different rewritten form.

This makes the variable elimination ordered based on x/y instead of (not
x)/(not y).

This is not a general solution but fixes ~5 proof checking failures on
SMT-LIB BV.

Fixes https://github.com/cvc5/cvc5-projects/issues/760.