Avoid deadlock in HO model construction (#11723)
This avoids a deadlock in HO model construction which was leading to
assertion failures that models failed to build.

The issue was that all terms were either considered assignable (legal to
assign a value in the model if not already assigned), or evaluable
(waiting for its children to have values so that its value was the
result of its evaluation).

In HO model construction, function variables are not considered
"assignable" since we assign their values based on the `APPLY_UF` that
involve them. However, they are also not "evaluable" since they have no
children.

This checks for this case.

Note that our behavior became more strict recently
https://github.com/cvc5/cvc5/pull/11322, which led to this failure.