Use arith substitution context for case of congruence manager conflicts (#11708)
This fixes a very rare proof hole that occurs on one benchmark from
SMT-LIB, which appears to be due to non-determinism in the arrays
rewriter. This benchmark was safe-guarded in
https://github.com/cvc5/cvc5/pull/11688 and added to regress3 (ddsmt
cannot minimize it in a reasonable amount of time). That PR fixed the
proof checking failure but left a hole of kind MACRO_SR_PRED_TRANSFORM.

This PR patches the current issue by simplifying proof production for
one of the arithmetic congruence conflict cases. In particular, we know
use a more "controlled" substitution context, ArithSubsTermContext, to
justify the rewrite, which should not be dependent on how the resulting
monomials are rewritten. It also should lead to smaller proofs for these
cases.

The same utility introduced in this PR (ArithSubsTermContext) will also
be used on https://github.com/cvc5/cvc5/pull/11453.