More minor changes to BV rewriter for proofs (#11704)
In particular we use explicit repeat terms in a few places to match RARE
rules.

The mult-slt-mult rewrite is simplified slightly to avoid more matching
cases. In particular we now return `(bvsub y x)` instead of `t` where
`y` is either `(bvadd x t)` or `(bvadd t x)`.

This was +125-110 on SMT-LIB logics with bitvectors, all differences
close to timeout.