Add fine grained proof support for quantifiers variable eliminations (#11698)
This is necessary to ensure the remaining variable eliminations succeed
in proof elaboration. It makes the proof elaboration for
MACRO_QUANT_VAR_ELIM_EQ pass a CDProof to the quantifiers rewriter to
provide fine grained steps for which we know proof elaboration will
succeed.

In particular this makes the quantifiers rewriter call the Boolean
rewriter for MACRO_BOOL_BV_INVERT_SOLVE, which requires a custom macro
rewrite and elaboration.