Do subproof merging when all processing RARE rewrites (#11820)
This is a performance optimization for proof generation.

This makes a single run of a proof updater process all RARE rewrites in
a single pass. This has the advantage that shared subgoals are only
proven by the RARE reconstruction once.

This adds a new callback to the proof node updater callback "finalize"
which is a clearer interface for when we are done processing a proof
node.