Catch all mixed arithmetic cases with --strict-parsing (#11817)
Addresses the inconsistency on
https://github.com/cvc5/cvc5/issues/11004.

This extends `--strict-parsing` to catch errors due to the use of mixed
arithmetic.

The original benchmark on https://github.com/cvc5/cvc5/issues/11004
already fails with `--strict-parsing` due to several non-compliant uses
of mixed arithmetic (e.g. equality between Int and Real). This extends
these checks so that we catch all non-compliant uses of arithmetic,
including `to_real` applied to a Real, as demonstrated by the minimal
example on that issue.

We continue to support the (non-compliant) extension of `to_real`
applied to reals when strict parsing is disabled (by default), which we
interpret as a no-op. This extension disagrees with z3, which appears to
insert a casting to int.