Make proof constructor for strings more robust (#11828)
This fixes a proof hole that was encountered in a run of SMT-LIB with
`--decision=internal` and was previously added as a regression
(regress0/strings/dd_str_term_529.smt2).

It changes the strings inference proof constructor to use a more
targetted method for proving equivalence between expected/actual
conclusions for a variable splitting lemma.