Optimizations for RARE reconstruction (#11829)
This fixes an issue with how failures were stored. In particular, we
were not properly caching whether the failure was independent of depth.

This furthermore introduces an optimization where we track whether the
current run failed due to resource limits globally. If not, then we know
that increasing the recursion limit will not have an impact and hence we
fail early.

It also corrects an issue where with the latest changes we would fail to
prove rewrites involving Boolean variables, e.g. `A` is equal to `A =
true`.

This improves overall end-to-end time for generating+checking proofs by
roughly 5% in my latest tests.