Introduced non-closed converter, more robust guard for non-closed terms in model blocker (#11854)
Introduces a node converter to eliminate terms that are not "closed",
i.e. may appear in models but can't be reasserted to the solver. This
will be used e.g. for model-based quantifier instantiation in a later
PR.

This utility can also be used to make the model blocker more robust,
e.g. to not introduce witness terms for certain string models.

Fixes https://github.com/cvc5/cvc5-projects/issues/766.

---------

Co-authored-by: Daniel Larraz <daniel-larraz@users.noreply.github.com>