Fix equalities in bitvector ppStaticRewrite (#11862)
The BV solver had cases to rewrite equalities in ppRewrite, which is not
called on equalities.

These rewrites are moved to the appropriate place (ppStaticRewrite).

The "bitwise equal" rewrites were enabled by default (despite being dead
code to the above mistake). Enabling them again led to worse performance
on SMT-LIB logics with BV (+268-425).

Moreover we do not have RARE rules that cover this case. This option is
thus demoted to "expert" and disabled by default.

Thus, this PR preserves behavior but ensures these rewrites are fired if
the user requests them.