Properly handle ambiguous constants in CPC (#11845)
This takes advantage of the fact that ethos now has SMT-LIB compliant
support for ordinary (non datatype constructor) ambiguous functions.

It updates the 5 constants in the signature to proper definitions.

In particular, we now will properly reference check SMT-LIB inputs with
set.empty / seq.empty.

We require refactoring definitions of `seq.empty` and `@seq.empty` to
account for this change, thus RARE rules now use the name `$seq_empty`
(which continues to refer to the program). The macro `$mk_emptystr` is
furthermore removed for simplicity.

This PR also moves reasoning about array constants to the expert
signature, as array constants are not available in safe mode.