[proofs] [alethe] Generalize legal type checking in Alethe backend (#11907)
Based on the robust test at #11876 for whether variables have types
supported in the Alethe backend. The previous test was not considering
when the unsupported types were component types (for example, the
children of an Array type) or in the subfields of datatypes (for
example, arguments of constructor types). Regressions exercising this
behavior are in #11876.