Update CPC to use eo::eq instead of eo::is_eq when applicable (#11892)
The new version of ethos now has `eo::eq` available, which is like
`eo::is_eq` but does not evaluate if one of the arguments is stuck.

We should prefer this operator whenever applicable, since it has better
guarantees.

This updates 113/119 uses of `eo::is_eq` to `eo::eq`. The remaining 6
*intentionally* make use of the `eo::is_eq` operator being total.